Minor change for hand attacking from harm mode

+++ Content.Shared/_Finster/Rulebook/SkillLevel.cs
@@ -3,7 +3,7 @@ using Robust.Shared.Serialization;
 namespace Content.Shared._Finster.Rulebook;
 
 [Serializable, NetSerializable]
-public enum SkillLevel : int
+public enum SkillLevel : byte
 {
     Weak,
     Basic,
